+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 
 ## master
 
+* Use the `cmdargs` package to parse CLI arguments.
 * Split library into `Console.BnbStaking` module hierarchy.
 
 ## v0.1.0.0